refactor: 전략패턴 시, 타입드 인터페이스 사용으로 변경
- 카드, 포인트 구현체에 따라 info의 구현체도 다른 부분 해결하기 위함.
- casting하기 전과 후를 모두 사용할 수 있는 부분 개선하기 위해서 수정
