+++ apps/commerce-api/src/main/java/com/loopers/application/payment/dto/PaymentInfo.java
@@ -11,7 +11,12 @@ public class PaymentInfo {
     // 공통 결제 정보
     @Getter
     @AllArgsConstructor(access = AccessLevel.PROTECTED)
-    public static abstract class Pay {
+    //PaymentInfo.Pay가 sealed이면 switch가 컴파일 타임에 누락 케이스를 잡아준다.
+    //permits 명시: Pay의 서브타입을 모두 permits에 나열해야 함.
+    //sealed 클래스는 abstract, final, non-sealed 중 하나여야 함.
+    //sealed 클래스로 안하고 usecase에서 switch 시에 default 처리 해도 된다.
+    public sealed static abstract class Pay
+            permits CardPay, PointPay{
         private final Long userId;
         private final Long orderId;
         private final PaymentMethod paymentMethod;
@@ -21,7 +26,7 @@ public class PaymentInfo {
     // 카드 결제 정보
     // =============================================================================
     @Getter
-    public static class CardPay extends Pay {
+    public static final class CardPay extends Pay {
         private final String cardNumber;
         private final String cardType;
 
@@ -50,14 +55,13 @@ public class PaymentInfo {
             );
 
         }
-
     }
 
     // =============================================================================
     // 포인트 결제 정보
     // =============================================================================
     @Getter
-    public static class PointPay extends Pay {
+    public static final class PointPay extends Pay {
 
         @Builder(access = AccessLevel.PRIVATE)
         private PointPay(Long userId, Long orderId) {
